src/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.9
django-celery-beat==2.2.1
django-feedparser==0.2.1
django-filter==21.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
dnspython==2.1.0
eventlet==0.32.0
feedparser==6.0.8
greenlet==1.1.2
idna==3.3
kombu==5.2.1
Markdown==3.3.4
numpy==1.19.5
pandas==1.1.5
prompt-toolkit==3.0.22
psycopg2==2.8.4
psycopg2-binary>=2.8
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5









